Made updates to company register api.
